# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameProfileMeta = { [key: string]: unknown };

export interface GameProfile {
  gameId: string;
  currency: number;
  meta: GameProfileMeta;
  updatedAt: string;
}

export type GameProfilePatchMeta = { [key: string]: unknown };

export interface GameProfilePatch {
  currency?: number;
  meta?: GameProfilePatchMeta;
}

export type MatchRecordMeta = { [key: string]: unknown };

export type MatchRecordMatchMeta = { [key: string]: unknown };

export interface MatchRecord {
  won: boolean;
  rewardGbux?: number;
  seed?: number;
  mode?: string;
  score?: number;
  meta?: MatchRecordMeta;
  matchMeta?: MatchRecordMatchMeta;
}

export interface MatchRecordResponse {
  ok: boolean;
  profile: GameProfile;
}