# lib/api-zod/src/generated/types/gameProfile.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameProfileMeta } from './gameProfileMeta';

export interface GameProfile {
  gameId: string;
  currency: number;
  meta: GameProfileMeta;
  updatedAt: string;
}