# lib/api-zod/src/generated/types/gameProfilePatch.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameProfilePatchMeta } from './gameProfilePatchMeta';

export interface GameProfilePatch {
  currency?: number;
  meta?: GameProfilePatchMeta;
}