# lib/api-zod/src/generated/types/matchRecord.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchRecordMatchMeta } from './matchRecordMatchMeta';
import type { MatchRecordMeta } from './matchRecordMeta';

export interface MatchRecord {
  won: boolean;
  rewardGbux?: number;
  seed?: number;
  mode?: string;
  score?: number;
  meta?: MatchRecordMeta;
  matchMeta?: MatchRecordMatchMeta;
}