# lib/api-zod/src/generated/types/matchRecordResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameProfile } from './gameProfile';

export interface MatchRecordResponse {
  ok: boolean;
  profile: GameProfile;
}